Extract ListBox classes
