+++ src/System.Windows.Forms/src/System/Windows/Forms/ListBox.cs
@@ -2566,1916 +2566,5 @@ namespace System.Windows.Forms
                     break;
             }
         }
-
-        /// <summary>
-        ///  This is similar to ArrayList except that it also
-        ///  mantains a bit-flag based state element for each item
-        ///  in the array.
-        ///
-        ///  The methods to enumerate, count and get data support
-        ///  virtualized indexes.  Indexes are virtualized according
-        ///  to the state mask passed in.  This allows ItemArray
-        ///  to be the backing store for one read-write "master"
-        ///  collection and serveral read-only collections based
-        ///  on masks.  ItemArray supports up to 31 masks.
-        /// </summary>
-        internal class ItemArray : IComparer
-        {
-            private static int lastMask = 1;
-
-            private readonly ListControl listControl;
-            private Entry[] entries;
-            private int count;
-            private int version;
-
-            public ItemArray(ListControl listControl)
-            {
-                this.listControl = listControl;
-            }
-
-            internal IReadOnlyList<Entry> Entries => entries;
-
-            /// <summary>
-            ///  The version of this array.  This number changes with each
-            ///  change to the item list.
-            /// </summary>
-            public int Version
-            {
-                get
-                {
-                    return version;
-                }
-            }
-
-            /// <summary>
-            ///  Adds the given item to the array.  The state is initially
-            ///  zero.
-            /// </summary>
-            public object Add(object item)
-            {
-                EnsureSpace(1);
-                version++;
-                entries[count] = new Entry(item);
-                return entries[count++];
-            }
-
-            /// <summary>
-            ///  Adds the given collection of items to the array.
-            /// </summary>
-            public void AddRange(ICollection items)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-                EnsureSpace(items.Count);
-                foreach (object i in items)
-                {
-                    entries[count++] = new Entry(i);
-                }
-                version++;
-            }
-
-            /// <summary>
-            ///  Clears this array.
-            /// </summary>
-            public void Clear()
-            {
-                if (count > 0)
-                {
-                    Array.Clear(entries, 0, count);
-                }
-
-                count = 0;
-                version++;
-            }
-
-            /// <summary>
-            ///  Allocates a new bitmask for use.
-            /// </summary>
-            public static int CreateMask()
-            {
-                int mask = lastMask;
-                lastMask <<= 1;
-                Debug.Assert(lastMask > mask, "We have overflowed our state mask.");
-                return mask;
-            }
-
-            /// <summary>
-            ///  Ensures that our internal array has space for
-            ///  the requested # of elements.
-            /// </summary>
-            private void EnsureSpace(int elements)
-            {
-                if (entries == null)
-                {
-                    entries = new Entry[Math.Max(elements, 4)];
-                }
-                else if (count + elements >= entries.Length)
-                {
-                    int newLength = Math.Max(entries.Length * 2, entries.Length + elements);
-                    Entry[] newEntries = new Entry[newLength];
-                    entries.CopyTo(newEntries, 0);
-                    entries = newEntries;
-                }
-            }
-
-            /// <summary>
-            ///  Turns a virtual index into an actual index.
-            /// </summary>
-            public int GetActualIndex(int virtualIndex, int stateMask)
-            {
-                if (stateMask == 0)
-                {
-                    return virtualIndex;
-                }
-
-                // More complex; we must compute this index.
-                int calcIndex = -1;
-                for (int i = 0; i < count; i++)
-                {
-                    if ((entries[i].state & stateMask) != 0)
-                    {
-                        calcIndex++;
-                        if (calcIndex == virtualIndex)
-                        {
-                            return i;
-                        }
-                    }
-                }
-
-                return -1;
-            }
-
-            /// <summary>
-            ///  Gets the count of items matching the given mask.
-            /// </summary>
-            public int GetCount(int stateMask)
-            {
-                // If mask is zero, then just give the main count
-                if (stateMask == 0)
-                {
-                    return count;
-                }
-
-                // more complex:  must provide a count of items
-                // based on a mask.
-
-                int filteredCount = 0;
-
-                for (int i = 0; i < count; i++)
-                {
-                    if ((entries[i].state & stateMask) != 0)
-                    {
-                        filteredCount++;
-                    }
-                }
-
-                return filteredCount;
-            }
-
-            /// <summary>
-            ///  Retrieves an enumerator that will enumerate based on
-            ///  the given mask.
-            /// </summary>
-            public IEnumerator GetEnumerator(int stateMask)
-            {
-                return GetEnumerator(stateMask, false);
-            }
-
-            /// <summary>
-            ///  Retrieves an enumerator that will enumerate based on
-            ///  the given mask.
-            /// </summary>
-            public IEnumerator GetEnumerator(int stateMask, bool anyBit)
-            {
-                return new EntryEnumerator(this, stateMask, anyBit);
-            }
-
-            /// <summary>
-            ///  Gets the item at the given index.  The index is
-            ///  virtualized against the given mask value.
-            /// </summary>
-            public object GetItem(int virtualIndex, int stateMask)
-            {
-                int actualIndex = GetActualIndex(virtualIndex, stateMask);
-
-                if (actualIndex == -1)
-                {
-                    throw new IndexOutOfRangeException();
-                }
-
-                return entries[actualIndex].item;
-            }
-            /// <summary>
-            ///  Gets the item at the given index.  The index is
-            ///  virtualized against the given mask value.
-            /// </summary>
-            internal object GetEntryObject(int virtualIndex, int stateMask)
-            {
-                int actualIndex = GetActualIndex(virtualIndex, stateMask);
-
-                if (actualIndex == -1)
-                {
-                    throw new IndexOutOfRangeException();
-                }
-
-                return entries[actualIndex];
-            }
-            /// <summary>
-            ///  Returns true if the requested state mask is set.
-            ///  The index is the actual index to the array.
-            /// </summary>
-            public bool GetState(int index, int stateMask)
-            {
-                return ((entries[index].state & stateMask) == stateMask);
-            }
-
-            /// <summary>
-            ///  Returns the virtual index of the item based on the
-            ///  state mask.
-            /// </summary>
-            public int IndexOf(object item, int stateMask)
-            {
-                int virtualIndex = -1;
-
-                for (int i = 0; i < count; i++)
-                {
-                    if (stateMask == 0 || (entries[i].state & stateMask) != 0)
-                    {
-                        virtualIndex++;
-                        if (entries[i].item.Equals(item))
-                        {
-                            return virtualIndex;
-                        }
-                    }
-                }
-
-                return -1;
-            }
-
-            /// <summary>
-            ///  Returns the virtual index of the item based on the
-            ///  state mask. Uses reference equality to identify the
-            ///  given object in the list.
-            /// </summary>
-            public int IndexOfIdentifier(object identifier, int stateMask)
-            {
-                int virtualIndex = -1;
-
-                for (int i = 0; i < count; i++)
-                {
-                    if (stateMask == 0 || (entries[i].state & stateMask) != 0)
-                    {
-                        virtualIndex++;
-                        if (entries[i] == identifier)
-                        {
-                            return virtualIndex;
-                        }
-                    }
-                }
-
-                return -1;
-            }
-
-            /// <summary>
-            ///  Inserts item at the given index.  The index
-            ///  is not virtualized.
-            /// </summary>
-            public void Insert(int index, object item)
-            {
-                EnsureSpace(1);
-
-                if (index < count)
-                {
-                    System.Array.Copy(entries, index, entries, index + 1, count - index);
-                }
-
-                entries[index] = new Entry(item);
-                count++;
-                version++;
-            }
-
-            /// <summary>
-            ///  Removes the given item from the array.  If
-            ///  the item is not in the array, this does nothing.
-            /// </summary>
-            public void Remove(object item)
-            {
-                int index = IndexOf(item, 0);
-
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            /// <summary>
-            ///  Removes the item at the given index.
-            /// </summary>
-            public void RemoveAt(int index)
-            {
-                count--;
-                for (int i = index; i < count; i++)
-                {
-                    entries[i] = entries[i + 1];
-                }
-                entries[count] = null;
-                version++;
-            }
-
-            /// <summary>
-            ///  Sets the item at the given index to a new value.
-            /// </summary>
-            public void SetItem(int index, object item)
-            {
-                entries[index].item = item;
-            }
-
-            /// <summary>
-            ///  Sets the state data for the given index.
-            /// </summary>
-            public void SetState(int index, int stateMask, bool value)
-            {
-                if (value)
-                {
-                    entries[index].state |= stateMask;
-                }
-                else
-                {
-                    entries[index].state &= ~stateMask;
-                }
-                version++;
-            }
-
-            /// <summary>
-            ///  Find element in sorted array. If element is not found returns a binary complement of index for inserting
-            /// </summary>
-            public int BinarySearch(object element)
-            {
-                return Array.BinarySearch(entries, 0, count, element, this);
-            }
-
-            /// <summary>
-            ///  Sorts our array.
-            /// </summary>
-            public void Sort()
-            {
-                Array.Sort(entries, 0, count, this);
-            }
-
-            public void Sort(Array externalArray)
-            {
-                Array.Sort(externalArray, this);
-            }
-
-            int IComparer.Compare(object item1, object item2)
-            {
-                if (item1 == null)
-                {
-                    if (item2 == null)
-                    {
-                        return 0; //both null, then they are equal
-                    }
-
-                    return -1; //item1 is null, but item2 is valid (greater)
-                }
-                if (item2 == null)
-                {
-                    return 1; //item2 is null, so item 1 is greater
-                }
-
-                if (item1 is Entry)
-                {
-                    item1 = ((Entry)item1).item;
-                }
-
-                if (item2 is Entry)
-                {
-                    item2 = ((Entry)item2).item;
-                }
-
-                string itemName1 = listControl.GetItemText(item1);
-                string itemName2 = listControl.GetItemText(item2);
-
-                CompareInfo compInfo = (Application.CurrentCulture).CompareInfo;
-                return compInfo.Compare(itemName1, itemName2, CompareOptions.StringSort);
-            }
-
-            /// <summary>
-            ///  This is a single entry in our item array.
-            /// </summary>
-            internal class Entry
-            {
-                public object item;
-                public int state;
-
-                public Entry(object item)
-                {
-                    this.item = item;
-                    state = 0;
-                }
-            }
-
-            /// <summary>
-            ///  EntryEnumerator is an enumerator that will enumerate over
-            ///  a given state mask.
-            /// </summary>
-            private class EntryEnumerator : IEnumerator
-            {
-                private readonly ItemArray items;
-                private readonly bool anyBit;
-                private readonly int state;
-                private int current;
-                private readonly int version;
-
-                /// <summary>
-                ///  Creates a new enumerator that will enumerate over the given state.
-                /// </summary>
-                public EntryEnumerator(ItemArray items, int state, bool anyBit)
-                {
-                    this.items = items;
-                    this.state = state;
-                    this.anyBit = anyBit;
-                    version = items.version;
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Moves to the next element, or returns false if at the end.
-                /// </summary>
-                bool IEnumerator.MoveNext()
-                {
-                    if (version != items.version)
-                    {
-                        throw new InvalidOperationException(SR.ListEnumVersionMismatch);
-                    }
-
-                    while (true)
-                    {
-                        if (current < items.count - 1)
-                        {
-                            current++;
-                            if (anyBit)
-                            {
-                                if ((items.entries[current].state & state) != 0)
-                                {
-                                    return true;
-                                }
-                            }
-                            else
-                            {
-                                if ((items.entries[current].state & state) == state)
-                                {
-                                    return true;
-                                }
-                            }
-                        }
-                        else
-                        {
-                            current = items.count;
-                            return false;
-                        }
-                    }
-                }
-
-                /// <summary>
-                ///  Resets the enumeration back to the beginning.
-                /// </summary>
-                void IEnumerator.Reset()
-                {
-                    if (version != items.version)
-                    {
-                        throw new InvalidOperationException(SR.ListEnumVersionMismatch);
-                    }
-
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Retrieves the current value in the enumerator.
-                /// </summary>
-                object IEnumerator.Current
-                {
-                    get
-                    {
-                        if (current == -1 || current == items.count)
-                        {
-                            throw new InvalidOperationException(SR.ListEnumCurrentOutOfRange);
-                        }
-
-                        return items.entries[current].item;
-                    }
-                }
-            }
-        }
-
-        // Items
-        /// <summary>
-            ///  A collection that stores objects.
-        /// </summary>
-        [ListBindable(false)]
-        public class ObjectCollection : IList
-        {
-            private readonly ListBox owner;
-            private ItemArray items;
-
-            public ObjectCollection(ListBox owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-                    ///  Initializes a new instance of ListBox.ObjectCollection based on another ListBox.ObjectCollection.
-                /// </summary>
-            public ObjectCollection(ListBox owner, ObjectCollection value)
-            {
-                this.owner = owner;
-                AddRange(value);
-            }
-
-            /// <summary>
-                    ///  Initializes a new instance of ListBox.ObjectCollection containing any array of objects.
-                /// </summary>
-            public ObjectCollection(ListBox owner, object[] value)
-            {
-                this.owner = owner;
-                AddRange(value);
-            }
-
-            /// <summary>
-            ///  Retrieves the number of items.
-            /// </summary>
-            public int Count
-            {
-                get
-                {
-                    return InnerArray.GetCount(0);
-                }
-            }
-
-            /// <summary>
-            ///  Internal access to the actual data store.
-            /// </summary>
-            internal ItemArray InnerArray
-            {
-                get
-                {
-                    if (items == null)
-                    {
-                        items = new ItemArray(owner);
-                    }
-                    return items;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Adds an item to the List box. For an unsorted List box, the item is
-            ///  added to the end of the existing list of items. For a sorted List box,
-            ///  the item is inserted into the list according to its sorted position.
-            ///  The item's toString() method is called to obtain the string that is
-            ///  displayed in the List box.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            public int Add(object item)
-            {
-                owner.CheckNoDataSource();
-                int index = AddInternal(item);
-                owner.UpdateHorizontalExtent();
-                return index;
-            }
-
-            private int AddInternal(object item)
-            {
-                if (item == null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-                int index = -1;
-                if (!owner.sorted)
-                {
-                    InnerArray.Add(item);
-                }
-                else
-                {
-                    if (Count > 0)
-                    {
-                        index = InnerArray.BinarySearch(item);
-                        if (index < 0)
-                        {
-                            index = ~index; // getting the index of the first element that is larger than the search value
-                                            //this index will be used for insert
-                        }
-                    }
-                    else
-                    {
-                        index = 0;
-                    }
-
-                    Debug.Assert(index >= 0 && index <= Count, "Wrong index for insert");
-                    InnerArray.Insert(index, item);
-                }
-                bool successful = false;
-
-                try
-                {
-                    if (owner.sorted)
-                    {
-                        if (owner.IsHandleCreated)
-                        {
-                            owner.NativeInsert(index, item);
-                            owner.UpdateMaxItemWidth(item, false);
-                            if (owner.selectedItems != null)
-                            {
-                                // Sorting may throw the LB contents and the selectedItem array out of synch.
-                                owner.selectedItems.Dirty();
-                            }
-                        }
-                    }
-                    else
-                    {
-                        index = Count - 1;
-                        if (owner.IsHandleCreated)
-                        {
-                            owner.NativeAdd(item);
-                            owner.UpdateMaxItemWidth(item, false);
-                        }
-                    }
-                    successful = true;
-                }
-                finally
-                {
-                    if (!successful)
-                    {
-                        InnerArray.Remove(item);
-                    }
-                }
-
-                return index;
-            }
-
-            int IList.Add(object item)
-            {
-                return Add(item);
-            }
-
-            public void AddRange(ObjectCollection value)
-            {
-                owner.CheckNoDataSource();
-                AddRangeInternal((ICollection)value);
-            }
-
-            public void AddRange(object[] items)
-            {
-                owner.CheckNoDataSource();
-                AddRangeInternal((ICollection)items);
-            }
-
-            internal void AddRangeInternal(ICollection items)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-                owner.BeginUpdate();
-                try
-                {
-                    foreach (object item in items)
-                    {
-                        // adding items one-by-one for performance
-                        // not using sort because after the array is sorted index of each newly added item will need to be found
-                        // AddInternal is based on BinarySearch and finds index without any additional cost
-                        AddInternal(item);
-                    }
-                }
-                finally
-                {
-                    owner.UpdateHorizontalExtent();
-                    owner.EndUpdate();
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the item with the specified index.
-            /// </summary>
-            [Browsable(false)]
-            [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-            public virtual object this[int index]
-            {
-                get
-                {
-                    if (index < 0 || index >= InnerArray.GetCount(0))
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    return InnerArray.GetItem(index, 0);
-                }
-                set
-                {
-                    owner.CheckNoDataSource();
-                    SetItemInternal(index, value);
-                }
-            }
-
-            /// <summary>
-            ///  Removes all items from the ListBox.
-            /// </summary>
-            public virtual void Clear()
-            {
-                owner.CheckNoDataSource();
-                ClearInternal();
-            }
-
-            /// <summary>
-            ///  Removes all items from the ListBox.  Bypasses the data source check.
-            /// </summary>
-            internal void ClearInternal()
-            {
-                //update the width.. to reset Scrollbars..
-                // Clear the selection state.
-                //
-                int cnt = owner.Items.Count;
-                for (int i = 0; i < cnt; i++)
-                {
-                    owner.UpdateMaxItemWidth(InnerArray.GetItem(i, 0), true);
-                }
-
-                if (owner.IsHandleCreated)
-                {
-                    owner.NativeClear();
-                }
-                InnerArray.Clear();
-                owner.maxWidth = -1;
-                owner.UpdateHorizontalExtent();
-            }
-
-            public bool Contains(object value)
-            {
-                return IndexOf(value) != -1;
-            }
-
-            /// <summary>
-            ///  Copies the ListBox Items collection to a destination array.
-            /// </summary>
-            public void CopyTo(object[] destination, int arrayIndex)
-            {
-                int count = InnerArray.GetCount(0);
-                for (int i = 0; i < count; i++)
-                {
-                    destination[i + arrayIndex] = InnerArray.GetItem(i, 0);
-                }
-            }
-
-            void ICollection.CopyTo(Array destination, int index)
-            {
-                int count = InnerArray.GetCount(0);
-                for (int i = 0; i < count; i++)
-                {
-                    destination.SetValue(InnerArray.GetItem(i, 0), i + index);
-                }
-            }
-
-            /// <summary>
-            ///  Returns an enumerator for the ListBox Items collection.
-            /// </summary>
-            public IEnumerator GetEnumerator()
-            {
-                return InnerArray.GetEnumerator(0);
-            }
-
-            public int IndexOf(object value)
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                return InnerArray.IndexOf(value, 0);
-            }
-
-            internal int IndexOfIdentifier(object value)
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                return InnerArray.IndexOfIdentifier(value, 0);
-            }
-
-            /// <summary>
-            ///  Adds an item to the List box. For an unsorted List box, the item is
-            ///  added to the end of the existing list of items. For a sorted List box,
-            ///  the item is inserted into the list according to its sorted position.
-            ///  The item's toString() method is called to obtain the string that is
-            ///  displayed in the List box.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            public void Insert(int index, object item)
-            {
-                owner.CheckNoDataSource();
-
-                if (index < 0 || index > InnerArray.GetCount(0))
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                if (item == null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-
-                // If the List box is sorted, then nust treat this like an add
-                // because we are going to twiddle the index anyway.
-                //
-                if (owner.sorted)
-                {
-                    Add(item);
-                }
-                else
-                {
-                    InnerArray.Insert(index, item);
-                    if (owner.IsHandleCreated)
-                    {
-                        bool successful = false;
-
-                        try
-                        {
-                            owner.NativeInsert(index, item);
-                            owner.UpdateMaxItemWidth(item, false);
-                            successful = true;
-                        }
-                        finally
-                        {
-                            if (!successful)
-                            {
-                                InnerArray.RemoveAt(index);
-                            }
-                        }
-                    }
-                }
-                owner.UpdateHorizontalExtent();
-            }
-
-            /// <summary>
-            ///  Removes the given item from the ListBox, provided that it is
-            ///  actually in the list.
-            /// </summary>
-            public void Remove(object value)
-            {
-                int index = InnerArray.IndexOf(value, 0);
-
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            /// <summary>
-            ///  Removes an item from the ListBox at the given index.
-            /// </summary>
-            public void RemoveAt(int index)
-            {
-                owner.CheckNoDataSource();
-
-                if (index < 0 || index >= InnerArray.GetCount(0))
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                owner.UpdateMaxItemWidth(InnerArray.GetItem(index, 0), true);
-
-                // Update InnerArray before calling NativeRemoveAt to ensure that when
-                // SelectedIndexChanged is raised (by NativeRemoveAt), InnerArray's state matches wrapped LB state.
-                InnerArray.RemoveAt(index);
-
-                if (owner.IsHandleCreated)
-                {
-                    owner.NativeRemoveAt(index);
-                }
-
-                owner.UpdateHorizontalExtent();
-            }
-
-            internal void SetItemInternal(int index, object value)
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                if (index < 0 || index >= InnerArray.GetCount(0))
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                owner.UpdateMaxItemWidth(InnerArray.GetItem(index, 0), true);
-                InnerArray.SetItem(index, value);
-
-                // If the native control has been created, and the display text of the new list item object
-                // is different to the current text in the native list item, recreate the native list item...
-                if (owner.IsHandleCreated)
-                {
-                    bool selected = (owner.SelectedIndex == index);
-                    if (string.Compare(owner.GetItemText(value), owner.NativeGetItemText(index), true, CultureInfo.CurrentCulture) != 0)
-                    {
-                        owner.NativeRemoveAt(index);
-                        owner.SelectedItems.SetSelected(index, false);
-                        owner.NativeInsert(index, value);
-                        owner.UpdateMaxItemWidth(value, false);
-                        if (selected)
-                        {
-                            owner.SelectedIndex = index;
-                        }
-                    }
-                    else
-                    {
-                        // FOR COMPATIBILITY REASONS
-                        if (selected)
-                        {
-                            owner.OnSelectedIndexChanged(EventArgs.Empty); //will fire selectedvaluechanged
-                        }
-                    }
-                }
-                owner.UpdateHorizontalExtent();
-            }
-        } // end ObjectCollection
-
-        //******************************************************************************************
-        // IntegerCollection
-        public class IntegerCollection : IList
-        {
-            private readonly ListBox owner;
-            private int[] innerArray;
-            private int count = 0;
-
-            public IntegerCollection(ListBox owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-            ///  Number of current selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    return count;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool Contains(int item)
-            {
-                return IndexOf(item) != -1;
-            }
-
-            bool IList.Contains(object item)
-            {
-                if (item is int)
-                {
-                    return Contains((int)item);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public void Clear()
-            {
-                count = 0;
-                innerArray = null;
-            }
-
-            public int IndexOf(int item)
-            {
-                int index = -1;
-
-                if (innerArray != null)
-                {
-                    index = Array.IndexOf(innerArray, item);
-
-                    // We initialize innerArray with more elements than needed in the method EnsureSpace,
-                    // and we don't actually remove element from innerArray in the method RemoveAt,
-                    // so there maybe some elements which are not actually in innerArray will be found
-                    // and we need to filter them out
-                    if (index >= count)
-                    {
-                        index = -1;
-                    }
-                }
-
-                return index;
-            }
-
-            int IList.IndexOf(object item)
-            {
-                if (item is int)
-                {
-                    return IndexOf((int)item);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            /// <summary>
-            ///  Add a unique integer to the collection in sorted order.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            private int AddInternal(int item)
-            {
-                EnsureSpace(1);
-
-                int index = IndexOf(item);
-                if (index == -1)
-                {
-                    innerArray[count++] = item;
-                    Array.Sort(innerArray, 0, count);
-                    index = IndexOf(item);
-                }
-                return index;
-            }
-
-            /// <summary>
-            ///  Adds a unique integer to the collection in sorted order.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            public int Add(int item)
-            {
-                int index = AddInternal(item);
-                owner.UpdateCustomTabOffsets();
-
-                return index;
-            }
-
-            int IList.Add(object item)
-            {
-                if (!(item is int))
-                {
-                    throw new ArgumentException(nameof(item));
-                }
-                return Add((int)item);
-            }
-
-            public void AddRange(int[] items)
-            {
-                AddRangeInternal((ICollection)items);
-            }
-
-            public void AddRange(IntegerCollection value)
-            {
-                AddRangeInternal((ICollection)value);
-            }
-
-            /// <summary>
-            ///  Add range that bypasses the data source check.
-            /// </summary>
-            private void AddRangeInternal(ICollection items)
-            {
-                if (items == null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-                owner.BeginUpdate();
-                try
-                {
-                    EnsureSpace(items.Count);
-                    foreach (object item in items)
-                    {
-                        if (!(item is int))
-                        {
-                            throw new ArgumentException(nameof(item));
-                        }
-                        else
-                        {
-                            AddInternal((int)item);
-                        }
-                    }
-                    owner.UpdateCustomTabOffsets();
-                }
-                finally
-                {
-                    owner.EndUpdate();
-                }
-            }
-
-            /// <summary>
-            ///  Ensures that our internal array has space for
-            ///  the requested # of elements.
-            /// </summary>
-            private void EnsureSpace(int elements)
-            {
-                if (innerArray == null)
-                {
-                    innerArray = new int[Math.Max(elements, 4)];
-                }
-                else if (count + elements >= innerArray.Length)
-                {
-                    int newLength = Math.Max(innerArray.Length * 2, innerArray.Length + elements);
-                    int[] newEntries = new int[newLength];
-                    innerArray.CopyTo(newEntries, 0);
-                    innerArray = newEntries;
-                }
-            }
-
-            void IList.Clear()
-            {
-                Clear();
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException(SR.ListBoxCantInsertIntoIntegerCollection);
-            }
-
-            void IList.Remove(object value)
-            {
-                if (!(value is int))
-                {
-                    throw new ArgumentException(nameof(value));
-                }
-                Remove((int)value);
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                RemoveAt(index);
-            }
-
-            /// <summary>
-            ///  Removes the given item from the array.  If
-            ///  the item is not in the array, this does nothing.
-            /// </summary>
-            public void Remove(int item)
-            {
-                int index = IndexOf(item);
-
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            /// <summary>
-            ///  Removes the item at the given index.
-            /// </summary>
-            public void RemoveAt(int index)
-            {
-                if (index < 0 || index >= count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                count--;
-                for (int i = index; i < count; i++)
-                {
-                    innerArray[i] = innerArray[i + 1];
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the specified selected item.
-            /// </summary>
-            public int this[int index]
-            {
-                get
-                {
-                    return innerArray[index];
-                }
-                set
-                {
-                    if (index < 0 || index >= count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-                    innerArray[index] = (int)value;
-                    owner.UpdateCustomTabOffsets();
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    if (!(value is int))
-                    {
-                        throw new ArgumentException(nameof(value));
-                    }
-                    else
-                    {
-                        this[index] = (int)value;
-                    }
-                }
-            }
-
-            public void CopyTo(Array destination, int index)
-            {
-                int cnt = Count;
-                for (int i = 0; i < cnt; i++)
-                {
-                    destination.SetValue(this[i], i + index);
-                }
-            }
-
-            IEnumerator IEnumerable.GetEnumerator()
-            {
-                return new CustomTabOffsetsEnumerator(this);
-            }
-
-            /// <summary>
-            ///  EntryEnumerator is an enumerator that will enumerate over
-            ///  a given state mask.
-            /// </summary>
-            private class CustomTabOffsetsEnumerator : IEnumerator
-            {
-                private readonly IntegerCollection items;
-                private int current;
-
-                /// <summary>
-                ///  Creates a new enumerator that will enumerate over the given state.
-                /// </summary>
-                public CustomTabOffsetsEnumerator(IntegerCollection items)
-                {
-                    this.items = items;
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Moves to the next element, or returns false if at the end.
-                /// </summary>
-                bool IEnumerator.MoveNext()
-                {
-                    if (current < items.Count - 1)
-                    {
-                        current++;
-                        return true;
-                    }
-                    else
-                    {
-                        current = items.Count;
-                        return false;
-                    }
-                }
-
-                /// <summary>
-                ///  Resets the enumeration back to the beginning.
-                /// </summary>
-                void IEnumerator.Reset()
-                {
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Retrieves the current value in the enumerator.
-                /// </summary>
-                object IEnumerator.Current
-                {
-                    get
-                    {
-                        if (current == -1 || current == items.Count)
-                        {
-                            throw new InvalidOperationException(SR.ListEnumCurrentOutOfRange);
-                        }
-
-                        return items[current];
-                    }
-                }
-            }
-        }
-
-        //******************************************************************************************
-
-        // SelectedIndices
-        public class SelectedIndexCollection : IList
-        {
-            private readonly ListBox owner;
-
-            /* C#r: protected */
-            public SelectedIndexCollection(ListBox owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-            ///  Number of current selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    return owner.SelectedItems.Count;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool Contains(int selectedIndex)
-            {
-                return IndexOf(selectedIndex) != -1;
-            }
-
-            bool IList.Contains(object selectedIndex)
-            {
-                if (selectedIndex is int)
-                {
-                    return Contains((int)selectedIndex);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public int IndexOf(int selectedIndex)
-            {
-                // Just what does this do?  The selectedIndex parameter above is the index into the
-                // main object collection.  We look at the state of that item, and if the state indicates
-                // that it is selected, we get back the virtualized index into this collection.  Indexes on
-                // this collection match those on the SelectedObjectCollection.
-                if (selectedIndex >= 0 &&
-                    selectedIndex < InnerArray.GetCount(0) &&
-                    InnerArray.GetState(selectedIndex, SelectedObjectCollection.SelectedObjectMask))
-                {
-                    return InnerArray.IndexOf(InnerArray.GetItem(selectedIndex, 0), SelectedObjectCollection.SelectedObjectMask);
-                }
-
-                return -1;
-            }
-
-            int IList.IndexOf(object selectedIndex)
-            {
-                if (selectedIndex is int)
-                {
-                    return IndexOf((int)selectedIndex);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-            }
-
-            void IList.Clear()
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-            }
-
-            void IList.Remove(object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-            }
-
-            /// <summary>
-            ///  Retrieves the specified selected item.
-            /// </summary>
-            public int this[int index]
-            {
-                get
-                {
-                    object identifier = InnerArray.GetEntryObject(index, SelectedObjectCollection.SelectedObjectMask);
-                    return InnerArray.IndexOfIdentifier(identifier, 0);
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    throw new NotSupportedException(SR.ListBoxSelectedIndexCollectionIsReadOnly);
-                }
-            }
-
-            /// <summary>
-            ///  This is the item array that stores our data.  We share this backing store
-            ///  with the main object collection.
-            /// </summary>
-            private ItemArray InnerArray
-            {
-                get
-                {
-                    owner.SelectedItems.EnsureUpToDate();
-                    return ((ObjectCollection)owner.Items).InnerArray;
-                }
-            }
-
-            public void CopyTo(Array destination, int index)
-            {
-                int cnt = Count;
-                for (int i = 0; i < cnt; i++)
-                {
-                    destination.SetValue(this[i], i + index);
-                }
-            }
-
-            public void Clear()
-            {
-                if (owner != null)
-                {
-                    owner.ClearSelected();
-                }
-            }
-
-            public void Add(int index)
-            {
-                if (owner != null)
-                {
-                    ObjectCollection items = owner.Items;
-                    if (items != null)
-                    {
-                        if (index != -1 && !Contains(index))
-                        {
-                            owner.SetSelected(index, true);
-                        }
-                    }
-                }
-            }
-
-            public void Remove(int index)
-            {
-                if (owner != null)
-                {
-                    ObjectCollection items = owner.Items;
-                    if (items != null)
-                    {
-                        if (index != -1 && Contains(index))
-                        {
-                            owner.SetSelected(index, false);
-                        }
-                    }
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                return new SelectedIndexEnumerator(this);
-            }
-
-            /// <summary>
-            ///  EntryEnumerator is an enumerator that will enumerate over
-            ///  a given state mask.
-            /// </summary>
-            private class SelectedIndexEnumerator : IEnumerator
-            {
-                private readonly SelectedIndexCollection items;
-                private int current;
-
-                /// <summary>
-                ///  Creates a new enumerator that will enumerate over the given state.
-                /// </summary>
-                public SelectedIndexEnumerator(SelectedIndexCollection items)
-                {
-                    this.items = items;
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Moves to the next element, or returns false if at the end.
-                /// </summary>
-                bool IEnumerator.MoveNext()
-                {
-                    if (current < items.Count - 1)
-                    {
-                        current++;
-                        return true;
-                    }
-                    else
-                    {
-                        current = items.Count;
-                        return false;
-                    }
-                }
-
-                /// <summary>
-                ///  Resets the enumeration back to the beginning.
-                /// </summary>
-                void IEnumerator.Reset()
-                {
-                    current = -1;
-                }
-
-                /// <summary>
-                ///  Retrieves the current value in the enumerator.
-                /// </summary>
-                object IEnumerator.Current
-                {
-                    get
-                    {
-                        if (current == -1 || current == items.Count)
-                        {
-                            throw new InvalidOperationException(SR.ListEnumCurrentOutOfRange);
-                        }
-
-                        return items[current];
-                    }
-                }
-            }
-        }
-
-        // Should be "ObjectCollection", except we already have one of those.
-        public class SelectedObjectCollection : IList
-        {
-            // This is the bitmask used within ItemArray to identify selected objects.
-            internal static int SelectedObjectMask = ItemArray.CreateMask();
-
-            private readonly ListBox owner;
-            private bool stateDirty;
-            private int lastVersion;
-            private int count;
-
-            /* C#r: protected */
-            public SelectedObjectCollection(ListBox owner)
-            {
-                this.owner = owner;
-                stateDirty = true;
-                lastVersion = -1;
-            }
-
-            /// <summary>
-            ///  Number of current selected items.
-            /// </summary>
-            public int Count
-            {
-                get
-                {
-                    if (owner.IsHandleCreated)
-                    {
-                        SelectionMode current = (owner.selectionModeChanging) ? owner.cachedSelectionMode : owner.selectionMode;
-                        switch (current)
-                        {
-                            case SelectionMode.None:
-                                return 0;
-
-                            case SelectionMode.One:
-                                int index = owner.SelectedIndex;
-                                if (index >= 0)
-                                {
-                                    return 1;
-                                }
-                                return 0;
-
-                            case SelectionMode.MultiSimple:
-                            case SelectionMode.MultiExtended:
-                                return unchecked((int)(long)SendMessageW(owner, (WM)LB.GETSELCOUNT));
-                        }
-
-                        return 0;
-                    }
-
-                    // If the handle hasn't been created, we must do this the hard way.
-                    // Getting the count when using a mask is expensive, so cache it.
-                    //
-                    if (lastVersion != InnerArray.Version)
-                    {
-                        lastVersion = InnerArray.Version;
-                        count = InnerArray.GetCount(SelectedObjectMask);
-                    }
-
-                    return count;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            /// <summary>
-            ///  Called by the list box to dirty the selected item state.
-            /// </summary>
-            internal void Dirty()
-            {
-                stateDirty = true;
-            }
-
-            /// <summary>
-            ///  This is the item array that stores our data.  We share this backing store
-            ///  with the main object collection.
-            /// </summary>
-            private ItemArray InnerArray
-            {
-                get
-                {
-                    EnsureUpToDate();
-                    return ((ObjectCollection)owner.Items).InnerArray;
-                }
-            }
-
-            /// <summary>
-            ///  This is the function that Ensures that the selections are uptodate with
-            ///  current listbox handle selections.
-            /// </summary>
-            internal void EnsureUpToDate()
-            {
-                if (stateDirty)
-                {
-                    stateDirty = false;
-                    if (owner.IsHandleCreated)
-                    {
-                        owner.NativeUpdateSelection();
-                    }
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool Contains(object selectedObject)
-            {
-                return IndexOf(selectedObject) != -1;
-            }
-
-            public int IndexOf(object selectedObject)
-            {
-                return InnerArray.IndexOf(selectedObject, SelectedObjectMask);
-            }
-
-            int IList.Add(object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-            }
-
-            void IList.Clear()
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-            }
-
-            void IList.Remove(object value)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-            }
-
-            // A new internal method used in SelectedIndex getter...
-            // For a Multi select ListBox there can be two items with the same name ...
-            // and hence a object comparison is required...
-            // This method returns the "object" at the passed index rather than the "item" ...
-            // this "object" is then compared in the IndexOf( ) method of the itemsCollection.
-            //
-            internal object GetObjectAt(int index)
-            {
-                return InnerArray.GetEntryObject(index, SelectedObjectCollection.SelectedObjectMask);
-            }
-
-            /// <summary>
-            ///  Retrieves the specified selected item.
-            /// </summary>
-            [Browsable(false)]
-            [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-            public object this[int index]
-            {
-                get
-                {
-                    return InnerArray.GetItem(index, SelectedObjectMask);
-                }
-                set
-                {
-                    throw new NotSupportedException(SR.ListBoxSelectedObjectCollectionIsReadOnly);
-                }
-            }
-
-            public void CopyTo(Array destination, int index)
-            {
-                int cnt = InnerArray.GetCount(SelectedObjectMask);
-                for (int i = 0; i < cnt; i++)
-                {
-                    destination.SetValue(InnerArray.GetItem(i, SelectedObjectMask), i + index);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                return InnerArray.GetEnumerator(SelectedObjectMask);
-            }
-
-            /// <summary>
-            ///  This method returns if the actual item index is selected.  The index is the index to the MAIN
-            ///  collection, not this one.
-            /// </summary>
-            internal bool GetSelected(int index)
-            {
-                return InnerArray.GetState(index, SelectedObjectMask);
-            }
-
-            // when SelectedObjectsCollection::ItemArray is accessed we push the selection from Native ListBox into our .Net ListBox - see EnsureUpToDate()
-            // when we create the handle we need to be able to do the opposite : push the selection from .Net ListBox into Native ListBox
-            internal void PushSelectionIntoNativeListBox(int index)
-            {
-                // we can't use ItemArray accessor because this will wipe out our Selection collection
-                bool selected = ((ObjectCollection)owner.Items).InnerArray.GetState(index, SelectedObjectMask);
-                // push selection only if the item is actually selected
-                // this also takes care of the case where owner.SelectionMode == SelectionMode.One
-                if (selected)
-                {
-                    owner.NativeSetSelected(index, true /*we signal selection to the native listBox only if the item is actually selected*/);
-                }
-            }
-
-            /// <summary>
-            ///  Same thing for GetSelected.
-            /// </summary>
-            internal void SetSelected(int index, bool value)
-            {
-                InnerArray.SetState(index, SelectedObjectMask, value);
-            }
-
-            public void Clear()
-            {
-                if (owner != null)
-                {
-                    owner.ClearSelected();
-                }
-            }
-
-            public void Add(object value)
-            {
-                if (owner != null)
-                {
-                    ObjectCollection items = owner.Items;
-                    if (items != null && value != null)
-                    {
-                        int index = items.IndexOf(value);
-                        if (index != -1 && !GetSelected(index))
-                        {
-                            owner.SelectedIndex = index;
-                        }
-                    }
-                }
-            }
-
-            public void Remove(object value)
-            {
-                if (owner != null)
-                {
-                    ObjectCollection items = owner.Items;
-                    if (items != null & value != null)
-                    {
-                        int index = items.IndexOf(value);
-                        if (index != -1 && GetSelected(index))
-                        {
-                            owner.SetSelected(index, false);
-                        }
-                    }
-                }
-            }
-        }
     }
 }
